Conteúdo do app atualizado para o proposto pela Dra. Simone. Respostas do jogo do porquê fui eu que forneci :)

+++ IntervencaoNarrativa/Categories/NSString+LocalizedString.h
@@ -15,6 +15,6 @@
 - (NSString *)localizedStringInNamespace:(NSString *)nameSpace;
 
 - (NSString *)localizedStringWithParam:(id)param;
-- (NSString *)localizedStringWithIntParam:(int)param;
+- (NSString *)localizedStringWithIntParam:(NSInteger)param;
 
 @end
